pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multimodal-medical-ai"
version = "0.1.0"
description = "Multimodal medical exam discrepancy detection platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "celery[redis]>=5.4.0",
  "redis>=5.0.7",
  "sqlalchemy>=2.0.31",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.2",
  "PyJWT>=2.9.0",
  "mlflow>=2.15.0",
  "torch>=2.3.0",
  "transformers>=4.42.0",
  "opencv-python>=4.10.0.84",
  "openai-whisper>=20231117",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
